requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
httpx==0.28.1
python-multipart==0.0.18
itsdangerous==2.2.0
starlette==0.41.3
google-generativeai==0.8.3
